MAINT: Update for deprecations
Update for future deprecations

+++ randomgen/tests/test_against_numpy.py
@@ -51,7 +51,7 @@ def compare_1_input(f1, f2, is_small=False):
 def compare_2_input(f1, f2, is_np=False, is_scalar=False):
     if is_np:
         a, b = 10, 0.3
-        dtype = np.int
+        dtype = int
     else:
         a, b = 2, 3
         dtype = np.double
@@ -95,17 +95,17 @@ def compare_3_input(f1, f2, is_np=False):
         ((a, np.array([b] * 10), c), {"size": 10}),
         (
             (
-                np.ones((1, 37), dtype=np.int) * a,
-                np.ones((23, 1), dtype=np.int) * [b],
-                c * np.ones((7, 1, 1), dtype=np.int),
+                np.ones((1, 37), dtype=int) * a,
+                np.ones((23, 1), dtype=int) * [b],
+                c * np.ones((7, 1, 1), dtype=int),
             ),
             {},
         ),
         (
             (
-                np.ones((1, 37), dtype=np.int) * a,
-                np.ones((23, 1), dtype=np.int) * [b],
-                c * np.ones((7, 1, 1), dtype=np.int),
+                np.ones((1, 37), dtype=int) * a,
+                np.ones((23, 1), dtype=int) * [b],
+                c * np.ones((7, 1, 1), dtype=int),
             ),
             {"size": (7, 23, 37)},
         ),
